+++ admin/service/system/sys_authority.go
@@ -50,7 +50,7 @@ func (authorityService *AuthorityService) CopyAuthority(copyInfo response.SysAut
 	var baseMenu []system.SysBaseMenu
 	for _, v := range menus {
 		intNum, _ := strconv.Atoi(v.MenuId)
-		v.SysBaseMenu.ID = uint(intNum)
+		v.SysBaseMenu.ID = uint64(intNum)
 		baseMenu = append(baseMenu, v.SysBaseMenu)
 	}
 	copyInfo.Authority.SysBaseMenus = baseMenu
